+++ src/public-api.ts
@@ -1,17 +1,15 @@
 /*
  * Public API Surface of error
  */
-// Class.
-export { ValidationError } from './lib/validation-error.class';
-// Interface.
-export { ErrorMessage } from './interface/error-message.interface';
-
-/*
- * Experimental
- */
 export {
   // Class.
-  MessageBuilder,
-  MessageFunctionBuilder,
-  MessageBuilderTemplate,
-} from './message-builder/';
+  CommonError,
+  Error,
+  Errors,
+  RangeError,
+  RangeErrors,
+  TypeError,
+  TypeErrors,
+  ValidationError,
+  ValidationErrors
+} from './lib';