+++ src/type/allowed-callback.type.ts
@@ -1,4 +1,0 @@
-/**
- * Allowed callback function names available for the `ValidationError`.
- */
-export type VEAllowedCallback = 'setFix' | 'setMessage' | 'setProblem' | 'setTemplate';